+++ apis/rust/node/src/lib.rs
@@ -1,8 +1,8 @@
-pub use daemon::{Event, EventStream};
+pub use daemon_connection::{Event, EventStream};
 pub use dora_core;
 pub use dora_core::message::{uhlc, Metadata, MetadataParameters};
 pub use flume::Receiver;
 pub use node::DoraNode;
 
-mod daemon;
+mod daemon_connection;
 mod node;
